+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 <br/>
 
 <p>
+<b>Version 1.2.44</b> - September 2021<br/>
+- NEW: Scene Node: you can now add a pause in the command rule list (example, turn on light, wait 4000 milliseconds, turn off light).<br/>
+</p>
+<p>
 <b>Version 1.2.43</b> - September 2021<br/>
 - Watchdog Node: fixed the msg.connectGateway = true not actually reconnecting.<br/>
 </p>